Data types

+++ quick_functions.py
@@ -11,10 +11,10 @@ class QuickFunctions:
             sum_arr = [0, 0, 0, 0]
             for row in rows:
                 if row[0] != "Card Name" and row[0] != "Deck Totals":
-                    sum_arr[0] += float(row[2])
-                    sum_arr[1] += float(row[3])
-                    sum_arr[2] += float(row[4])
-                    sum_arr[3] += float(row[5])
+                    sum_arr[0] += int(row[2])
+                    sum_arr[1] += int(row[3])
+                    sum_arr[2] += int(row[4])
+                    sum_arr[3] += int(row[5])
 
         # Append the sum as a new row and overwrite the file
         with open(csv_path, mode="w", newline="") as file:
